+++ src/rpc.rs
@@ -1,7 +1,7 @@
 use {
     crate::{
-        quic_solana::ConnectionCacheIdentity, solana::sanitize_transaction_support_check,
-        transactions::SendTransactionsPool,
+        blocking_services::AllowTxSigner, quic_solana::ConnectionCacheIdentity,
+        solana::sanitize_transaction_support_check, transactions::SendTransactionsPool,
     },
     anyhow::Context as _,
     futures::future::{ready, BoxFuture, FutureExt, TryFutureExt},
@@ -40,6 +40,7 @@ pub enum RpcServerType {
         rpc: String,
         proxy_sanitize_check: bool,
         proxy_preflight_check: bool,
+        policies: Arc<dyn AllowTxSigner + Send + Sync + 'static>,
     },
 }
 
@@ -106,6 +107,7 @@ impl RpcServer {
                 rpc,
                 proxy_sanitize_check,
                 proxy_preflight_check,
+                policies,
             } => {
                 use rpc_solana_like::RpcServer;
 
@@ -114,6 +116,7 @@ impl RpcServer {
                     rpc,
                     proxy_sanitize_check,
                     proxy_preflight_check,
+                    policies,
                 )
                 .await?;
 
@@ -137,6 +140,7 @@ impl RpcServer {
         rpc: String,
         proxy_sanitize_check: bool,
         proxy_preflight_check: bool,
+        policies: Arc<dyn AllowTxSigner + Send + Sync + 'static>,
     ) -> anyhow::Result<rpc_solana_like::RpcServerImpl> {
         let rpc = Arc::new(SolanaRpcClient::new(rpc));
         let sanitize_supported = sanitize_transaction_support_check(&rpc).await?;
@@ -146,6 +150,7 @@ impl RpcServer {
             rpc,
             proxy_sanitize_check: proxy_sanitize_check && sanitize_supported,
             proxy_preflight_check,
+            policies,
         })
     }
 
@@ -264,6 +269,7 @@ pub mod rpc_admin {
 pub mod rpc_solana_like {
     use {
         crate::{
+            blocking_services::AllowTxSigner, metrics,
             payload::RpcSendTransactionConfigWithBlockList, rpc::invalid_params,
             solana::decode_and_deserialize, transaction_handler::TransactionHandler,
             transactions::SendTransactionsPool,
@@ -271,6 +277,7 @@ pub mod rpc_solana_like {
         jsonrpsee::{
             core::{async_trait, RpcResult},
             proc_macros::rpc,
+            types::{error::INVALID_REQUEST_CODE, ErrorObject},
         },
         solana_client::nonblocking::rpc_client::RpcClient as SolanaRpcClient,
         solana_rpc_client_api::response::RpcVersionInfo,
@@ -299,6 +306,7 @@ pub mod rpc_solana_like {
         pub rpc: Arc<SolanaRpcClient>,
         pub proxy_sanitize_check: bool,
         pub proxy_preflight_check: bool,
+        pub policies: Arc<dyn AllowTxSigner + Send + Sync + 'static>,
     }
 
     impl RpcServerImpl {
@@ -348,6 +356,19 @@ pub mod rpc_solana_like {
                     .into_binary_encoding()
                     .ok_or_else(|| invalid_params("unsupported encoding"))?,
             )?;
+
+            if self
+                .policies
+                .allow_tx_signer(&config_with_blocklist.blocklist_pdas, &transaction)
+            {
+                metrics::jet::increment_banned_transactions_total();
+
+                return Err(ErrorObject::owned::<()>(
+                    INVALID_REQUEST_CODE,
+                    "Transaction signer is banned",
+                    None,
+                ));
+            }
 
             self.handle_internal_transaction(transaction, config_with_blocklist)
                 .await